+++ packages/vite-plugin-vui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vite-plugin-vui
 
+## 2.0.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.19.46
+
 ## 2.0.45
 
 ### Patch Changes